refactor: :pushpin: Implements table in dashboard page
